The check for funds is in the store. The button to create proxy is disabled if there are not enough funds.
